conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20211102.0=he9d5cce_0
aiohttp=3.8.3=py39h6c40b1e_0
aiosignal=1.2.0=pyhd3eb1b0_0
altair=5.0.1=py39hecd8cb5_0
appnope=0.1.2=py39hecd8cb5_1001
arrow-cpp=11.0.0=h9980dd5_2
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py39hecd8cb5_0
attrs=22.1.0=py39hecd8cb5_0
aws-c-common=0.4.57=hb1e8313_1
aws-c-event-stream=0.1.6=h23ab428_5
aws-checksums=0.1.9=hb1e8313_0
aws-sdk-cpp=1.8.185=he271ece_0
backcall=0.2.0=pyhd3eb1b0_0
blas=2.117=openblas
blas-devel=3.9.0=17_osx64_openblas
blinker=1.4=py39hecd8cb5_0
boost-cpp=1.73.0=hca72f7f_12
bottleneck=1.3.5=py39h67323c0_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
c-ares=1.19.1=h6c40b1e_0
ca-certificates=2023.08.22=hecd8cb5_0
cachetools=4.2.2=pyhd3eb1b0_0
cairo=1.16.0=h0ab9d94_1001
certifi=2023.7.22=py39hecd8cb5_0
cffi=1.15.1=py39hc55c11b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39hecd8cb5_0
comm=0.1.2=py39hecd8cb5_0
cryptography=41.0.2=py39h2093b18_0
debugpy=1.6.7=py39hcec6c5f_0
decorator=5.1.1=pyhd3eb1b0_0
eigen=3.4.0=h940c156_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.2.2=h97e5cf8_0
flask=2.2.2=py39hecd8cb5_0
fontconfig=2.13.94=h10f422b_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.3.3=py39h6c40b1e_0
gettext=0.21.0=he85b6c0_1
gflags=2.2.2=h0a44026_0
giflib=5.2.1=h6c40b1e_3
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.30=py39hecd8cb5_0
glib=2.69.1=h8346a28_1
glog=0.5.0=h23ab428_0
gmp=6.2.1=h2e338ed_0
gnutls=3.6.15=hed9c0bf_0
graphite2=1.3.14=he9d5cce_1
grpc-cpp=1.48.2=h3afe56f_0
gst-plugins-base=1.14.1=hcec6c5f_1
gstreamer=1.14.1=h6c40b1e_1
harfbuzz=4.3.0=hffc734d_1
hdf5=1.12.1=nompi_h2f0ef1a_100
icu=58.2=h0a44026_1000
idna=3.4=py39hecd8cb5_0
importlib-metadata=6.0.0=py39hecd8cb5_0
importlib_metadata=6.0.0=hd3eb1b0_0
intel-openmp=2023.1.0=ha357a0b_43547
ipykernel=6.25.0=py39h20db666_0
ipython=8.12.2=py39hecd8cb5_0
ipywidgets=8.0.4=py39hecd8cb5_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.18.1=py39hecd8cb5_1
jinja2=3.1.2=py39hecd8cb5_0
jpeg=9e=h6c40b1e_1
jsonschema=4.17.3=py39hecd8cb5_0
jupyter_client=8.1.0=py39hecd8cb5_0
jupyter_core=5.3.0=py39hecd8cb5_0
jupyterlab_widgets=3.0.5=py39hecd8cb5_0
krb5=1.20.1=hdba6334_1
lame=3.100=hb7f2c08_1003
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libblas=3.9.0=17_osx64_openblas
libboost=1.73.0=h3fa6bed_12
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcblas=3.9.0=17_osx64_openblas
libclang=12.0.0=default_hbc2896b_2
libcurl=8.2.1=ha585b31_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_0
libedit=3.1.20221030=h6c40b1e_0
libev=4.33=h9ed2024_1
libevent=2.1.12=h0a4fc7d_0
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=12_3_0_h97931a8_1
libgfortran5=12.3.0=hbd3c1fe_1
libiconv=1.16=hca72f7f_2
libidn2=2.3.4=h6c40b1e_0
liblapack=3.9.0=17_osx64_openblas
liblapacke=3.9.0=17_osx64_openblas
libllvm12=12.0.1=hd011deb_2
libnghttp2=1.52.0=h1c88b7d_1
libopenblas=0.3.23=openmp_h429af6e_0
libopus=1.3.1=hc929b4f_1
libpng=1.6.39=h6c40b1e_0
libpq=12.15=hdb2fb19_1
libprotobuf=3.20.3=hfff2838_0
libsodium=1.0.18=h1de35cc_0
libssh2=1.10.0=hdb2fb19_2
libtasn1=4.19.0=hb7f2c08_0
libthrift=0.15.0=h48f73ad_2
libtiff=4.5.1=hcec6c5f_0
libunistring=0.9.10=h0d85af4_0
libvpx=1.7.0=h378b8a2_0
libwebp=1.2.4=hf6ce154_1
libwebp-base=1.2.4=h6c40b1e_1
libxml2=2.10.4=h930c0e2_0
libxslt=1.1.37=h5d22bc9_0
llvm-openmp=16.0.6=hff08bdf_0
lz4-c=1.9.4=hcec6c5f_0
markdown-it-py=2.2.0=py39hecd8cb5_1
markupsafe=2.1.1=py39hca72f7f_0
matplotlib-inline=0.1.6=py39hecd8cb5_0
mdurl=0.1.0=py39hecd8cb5_0
mkl=2023.1.0=h8e150cf_43559
mkl-service=2.4.0=py39h6c40b1e_1
multidict=6.0.2=py39hca72f7f_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=py39hecd8cb5_0
nettle=3.7.3=h230ac6f_1
nspr=4.35=hea0b92c_0
nss=3.89.1=hcec6c5f_0
numexpr=2.8.4=py39h75c1baf_0
numpy=1.23.5=py39hdfa1d0c_0
openblas=0.3.23=openmp_hbefa662_0
opencv=4.6.0=py39h9991a6c_5
openh264=2.1.1=h8346a28_0
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1v=hca72f7f_0
orc=1.7.4=h995b336_1
packaging=23.1=py39hecd8cb5_0
pandas=2.0.3=py39h3ea8b11_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=he49afe7_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py39hcec6c5f_0
pip=23.2.1=py39hecd8cb5_0
pixman=0.38.0=h01d97ff_1003
platformdirs=3.10.0=py39hecd8cb5_0
prompt-toolkit=3.0.36=py39hecd8cb5_0
protobuf=3.20.3=py39hcec6c5f_0
psutil=5.9.0=py39hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=11.0.0=py39he65a03e_1
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.7.1=py39hecd8cb5_0
pygments=2.15.1=py39hecd8cb5_1
pympler=0.9=py_0
pyopenssl=23.2.0=py39hecd8cb5_0
pyrsistent=0.18.0=py39hca72f7f_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.2=h88f2d9e_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2022.7=py39hecd8cb5_0
pyyaml=6.0.1=py39hdc70f33_0
pyzmq=25.1.0=py39hcec6c5f_0
qt-main=5.15.2=h719ae48_7
qt-webengine=5.15.9=h90a370e_7
qtwebkit=5.212=hbfab81c_5
re2=2022.04.01=he9d5cce_0
readline=8.2=hca72f7f_0
requests=2.31.0=py39hecd8cb5_0
rich=13.3.5=py39hecd8cb5_0
setuptools=68.0.0=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=he9d5cce_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.24.1=pyhd8ed1ab_1
tbb=2021.8.0=ha357a0b_0
tenacity=8.2.2=py39hecd8cb5_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py39hecd8cb5_0
tornado=6.3.2=py39h6c40b1e_0
traitlets=5.7.1=py39hecd8cb5_0
typing-extensions=4.7.1=py39hecd8cb5_0
typing_extensions=4.7.1=py39hecd8cb5_0
tzdata=2023c=h04d1e81_0
tzlocal=2.1=py39hecd8cb5_1
urllib3=1.26.16=py39hecd8cb5_0
utf8proc=2.6.1=h9ed2024_0
validators=0.18.2=pyhd3eb1b0_0
watchdog=3.0.0=py39h82e4197_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.3=py39hecd8cb5_0
wheel=0.38.4=py39hecd8cb5_0
widgetsnbextension=4.0.5=py39hecd8cb5_0
x264=1!157.20191217=h1de35cc_0
xz=5.4.2=h6c40b1e_0
yaml=0.2.5=h0d85af4_2
yarl=1.8.1=py39hca72f7f_0
zeromq=4.3.4=h23ab428_0
zipp=3.11.0=py39hecd8cb5_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
